Changes to support MKR 1010 and Nano 33 IOT
Changes to support MKR 1010 and Nano 33 IOT
